+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased][unreleased]
 ### Changed
 - Match WordPress 6.8 "Add New" -> "Add" wording change for post type and taxonomy labels
+- Compare host and port for comparing valid hosts
 
 ## [4.0.4] - 2024-02-01
 ### Added